+++ testing/it/e2e-selenium/src/test/java/com/adobe/cq/email/core/components/it/seljup/util/components/segmentation/SegmentationEditDialog.java
@@ -1,0 +1,26 @@
+/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ ~ Copyright 2022 Adobe
+ ~
+ ~ Licensed under the Apache License, Version 2.0 (the "License");
+ ~ you may not use this file except in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~     http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing, software
+ ~ distributed under the License is distributed on an "AS IS" BASIS,
+ ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ ~ See the License for the specific language governing permissions and
+ ~ limitations under the License.
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
+package com.adobe.cq.email.core.components.it.seljup.util.components.segmentation;
+
+import com.adobe.cq.email.core.components.it.seljup.util.commons.SegmentationEditor;
+import com.adobe.cq.wcm.core.components.it.seljup.util.components.tabs.TabsEditDialog;
+
+public class SegmentationEditDialog extends TabsEditDialog {
+
+    public SegmentationEditor getChildrenEditor() {
+        return new SegmentationEditor();
+    }
+}
